refactor: delete istex-tdm to create ip-to-machine-name
