--- istex-tdm/.gitignore
@@ -1,2 +1,0 @@
-autorisation-abes.json
-inist-ip.json
